refactor: split audio functions from speech module
