+++ apps/mac/Stet/Core/Audio/AudioLevelSource.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol AudioLevelSource: Sendable {
+    func makeAudioLevelStream() async -> AsyncStream<Double>
+}
