+++ apps/mac/StetTests/Core/LogicPrimitiveTests.swift
@@ -6,28 +6,6 @@ import Testing
 @MainActor
 @Suite("Logic Primitives")
 struct LogicPrimitiveTests {
-    @Test func audioLevelBridgeEmitsToMultipleStreamsAndFinishes() async {
-        let bridge = AudioLevelBridge()
-        let streamA = bridge.makeStream()
-        let streamB = bridge.makeStream()
-
-        bridge.emit(0.25)
-        bridge.finish()
-
-        var iteratorA = streamA.makeAsyncIterator()
-        var iteratorB = streamB.makeAsyncIterator()
-
-        let valueA = await iteratorA.next()
-        let valueB = await iteratorB.next()
-        let endA = await iteratorA.next()
-        let endB = await iteratorB.next()
-
-        #expect(valueA == 0.25)
-        #expect(valueB == 0.25)
-        #expect(endA == nil)
-        #expect(endB == nil)
-    }
-
     @Test func textProcessingPipelineRunsStagesInOrder() async throws {
         let pipeline = TextProcessingPipeline(stages: [
             .init(name: "trim") { $0.trimmingCharacters(in: .whitespaces) },
